Reduced the octree file size by half again by making sure that the chunk positions were saved as ints and not floats. Now about 4.8 MB

+++ source/include/world/octree.h
@@ -13,7 +13,7 @@ public:
   BlockType type;               //Type
   std::vector<Octree> subTree;  //Sparse SubTree
   int depth = 4;                    //Current Depth
-  glm::vec3 pos;
+  glm::ivec3 pos;
 
   //Stuff that is set live!
   int index;                    //Current Index
